[expo-notifications] Bring back separate Firebase Messaging Service
