+++ android/expoview/src/main/java/host/exp/exponent/fcm/ExpoFcmMessagingService.kt
@@ -1,11 +1,12 @@
 package host.exp.exponent.fcm
 
 import android.annotation.SuppressLint
+import expo.modules.notifications.service.ExpoFirebaseMessagingService
 import expo.modules.notifications.service.NotificationsService
 import expo.modules.notifications.service.interfaces.FirebaseMessagingDelegate
 
 @SuppressLint("MissingFirebaseInstanceTokenRefresh")
-class ExpoFcmMessagingService : NotificationsService() {
+class ExpoFcmMessagingService : ExpoFirebaseMessagingService() {
   override val firebaseMessagingDelegate: FirebaseMessagingDelegate by lazy {
     ExpoFirebaseMessagingDelegate(this)
   }
