+++ packages/expo-notifications/android/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
 
   <application>
     <service
-      android:name=".service.NotificationsService"
+      android:name=".service.ExpoFirebaseMessagingService"
       android:exported="false">
       <intent-filter android:priority="-1">
         <action android:name="com.google.firebase.MESSAGING_EVENT" />
